rename: item to items
