+++ src/App.jsx
@@ -2,7 +2,7 @@ import { Routes, Route } from 'react-router';
 import { Header } from '@/components/Header';
 import { Footer } from '@/components/Footer';
 import { LandingPage } from './pages/LandingPage';
-import { ItemPage } from './pages/ItemPage';
+import { ItemsPage } from './pages/ItemsPage';
 
 function App() {
   return (
@@ -10,7 +10,7 @@ function App() {
       <Header />
       <Routes>
         <Route path="/" element={<LandingPage />} />
-        <Route path="/items" element={<ItemPage />} />
+        <Route path="/items" element={<ItemsPage />} />
       </Routes>
       <Footer />
     </>
