--- src/pages/ItemPage/index.js
@@ -1,1 +1,0 @@
-export { ItemPage } from './ItemPage';
