+++ src/pages/ItemsPage/index.js
@@ -1,0 +1,1 @@
+export { ItemsPage } from './ItemsPage';
